Extract manager for handling total coin
