+++ app/src/main/java/io/horizontalsystems/bankwallet/core/App.kt
@@ -17,6 +17,7 @@ import io.horizontalsystems.bankwallet.core.factories.EvmAccountManagerFactory
 import io.horizontalsystems.bankwallet.core.managers.*
 import io.horizontalsystems.bankwallet.core.providers.*
 import io.horizontalsystems.bankwallet.core.storage.*
+import io.horizontalsystems.bankwallet.modules.balance.TotalBalancePlatformCoinManager
 import io.horizontalsystems.bankwallet.modules.hsnft.HsNftApiProvider
 import io.horizontalsystems.bankwallet.modules.keystore.KeyStoreActivity
 import io.horizontalsystems.bankwallet.modules.launcher.LauncherActivity
@@ -101,6 +102,7 @@ class App : CoreApp(), WorkConfiguration.Provider {
         lateinit var evmBlockchainManager: EvmBlockchainManager
         lateinit var nftManager: NftManager
         lateinit var evmLabelManager: EvmLabelManager
+        lateinit var totalBalancePlatformCoinManager: TotalBalancePlatformCoinManager
     }
 
     override val testMode = BuildConfig.testMode
@@ -276,6 +278,8 @@ class App : CoreApp(), WorkConfiguration.Provider {
 
         wc2Service = WC2Service()
         wc2SessionManager = WC2SessionManager(accountManager, WC2SessionStorage(appDatabase), wc2Service, wc2Manager)
+
+        totalBalancePlatformCoinManager = TotalBalancePlatformCoinManager(coinManager, localStorage)
     }
 
     private fun initializeWalletConnectV2(appConfig: AppConfigProvider) {
